+++ chakra-ui-steps/src/components/Step/index.tsx
@@ -3,23 +3,23 @@ import {
   forwardRef,
   HTMLChakraProps,
   ThemingProps,
+  useStyles,
 } from '@chakra-ui/system';
-import { useMultiStyleConfig } from "@chakra-ui/react"
 import { Collapse } from '@chakra-ui/transition';
 import { dataAttr } from '@chakra-ui/utils';
 import { AnimatePresence } from 'framer-motion';
+import * as React from 'react';
 import { useStepsContext } from '../../context';
 import { Connector } from '../Connector';
 import { StepIcon } from '../StepIcon';
 import { StepLabel } from '../StepLabel';
-import React, { ComponentType, ReactNode, Ref, RefCallback, useCallback } from 'react';
 
 export interface StepProps extends HTMLChakraProps<'li'> {
-  label?: string | ReactNode;
+  label?: string | React.ReactNode;
   description?: string;
-  icon?: ComponentType<any>;
+  icon?: React.ComponentType<any>;
   state?: 'loading' | 'error';
-  checkIcon?: ComponentType<any>;
+  checkIcon?: React.ComponentType<any>;
   isCompletedStep?: boolean;
   isKeepError?: boolean;
 }
@@ -35,7 +35,7 @@ interface StepInternalConfig extends ThemingProps {
 interface FullStepProps extends StepProps, StepInternalConfig {}
 
 export const Step = forwardRef<StepProps, 'li'>(
-  (props, ref: Ref<any>) => {
+  (props, ref: React.Ref<any>) => {
     const {
       children,
       description,
@@ -50,7 +50,7 @@ export const Step = forwardRef<StepProps, 'li'>(
       label,
       ...styleProps
     } = props as FullStepProps;
-    
+
     const {
       isVertical,
       isError,
@@ -63,7 +63,7 @@ export const Step = forwardRef<StepProps, 'li'>(
       stepCount,
     } = useStepsContext();
 
-    const {step, stepContainer, stepIconContainer} = useMultiStyleConfig("Steps");
+    const { step, stepContainer, stepIconContainer } = useStyles();
 
     const hasVisited = isCurrentStep || isCompletedStep;
 
@@ -75,7 +75,7 @@ export const Step = forwardRef<StepProps, 'li'>(
       }
     };
 
-    const containerRef: RefCallback<HTMLDivElement> = useCallback(
+    const containerRef: React.RefCallback<HTMLDivElement> = React.useCallback(
       (node) => {
         if (node && setWidths) {
           setWidths((prev) => {
@@ -86,7 +86,7 @@ export const Step = forwardRef<StepProps, 'li'>(
           });
         }
       },
-      [stepIconContainer?.width, stepIconContainer?.height]
+      [stepIconContainer.width, stepIconContainer.height]
     );
 
     return (
@@ -97,10 +97,11 @@ export const Step = forwardRef<StepProps, 'li'>(
           aria-disabled={!hasVisited}
           __css={{
             opacity,
-            flexDirection: isVertical ? 'column' : 'row',
+            flexDir: isVertical ? 'column' : 'row',
             alignItems: isVertical || isLabelVertical ? 'flex-start' : 'center',
             flex: isLastStep && !isVertical ? '0 0 auto' : '1 0 auto',
-            justifyContent: isLastStep && !isVertical ? 'flex-end' : 'flex-start',
+            justifyContent:
+              isLastStep && !isVertical ? 'flex-end' : 'flex-start',
             _hover: {
               cursor: clickable ? 'pointer' : 'default',
             },
@@ -111,7 +112,7 @@ export const Step = forwardRef<StepProps, 'li'>(
           <chakra.div
             ref={containerRef}
             __css={{
-              flexDirection: isLabelVertical ? 'column' : 'row',
+              flexDir: isLabelVertical ? 'column' : 'row',
               ...stepContainer,
             }}
           >
@@ -129,7 +130,7 @@ export const Step = forwardRef<StepProps, 'li'>(
               data-highlighted={dataAttr(isCompletedStep)}
               data-clickable={dataAttr(clickable)}
             >
-              <AnimatePresence mode="wait">
+              <AnimatePresence exitBeforeEnter>
                 <StepIcon
                   {...{
                     index,